Move asset source related code from Media.Browser to Media
