+++ Neos.Media/Classes/Domain/Model/AssetSource/AssetProxy/SupportsIptcMetadata.php
@@ -1,8 +1,8 @@
 <?php
-namespace Neos\Media\Browser\AssetSource\AssetProxy;
+namespace Neos\Media\Domain\Model\AssetSource\AssetProxy;
 
 /*
- * This file is part of the Neos.Media.Browser package.
+ * This file is part of the Neos.Media package.
  *
  * (c) Contributors of the Neos Project - www.neos.io
   *
