+++ Neos.Media/Classes/Domain/Model/AssetSource/AssetProxyQueryResult.php
@@ -1,8 +1,8 @@
 <?php
-namespace Neos\Media\Browser\AssetSource;
+namespace Neos\Media\Domain\Model\AssetSource;
 
 /*
- * This file is part of the Neos.Media.Browser package.
+ * This file is part of the Neos.Media package.
  *
  * (c) Contributors of the Neos Project - www.neos.io
   *
@@ -11,7 +11,7 @@ namespace Neos\Media\Browser\AssetSource;
  * source code.
  */
 
-use Neos\Media\Browser\AssetSource\AssetProxy\AssetProxy;
+use Neos\Media\Domain\Model\AssetSource\AssetProxy\AssetProxy;
 
 interface AssetProxyQueryResult extends \Countable, \Iterator, \ArrayAccess
 {
