+++ Neos.Media/Classes/Domain/Model/AssetSource/AssetSourceAwareInterface.php
@@ -1,0 +1,38 @@
+<?php
+namespace Neos\Media\Domain\Model\AssetSource;
+
+/*
+ * This file is part of the Neos.Media package.
+ *
+ * (c) Contributors of the Neos Project - www.neos.io
+  *
+ * This package is Open Source Software. For the full copyright and license
+ * information, please view the LICENSE file which was distributed with this
+ * source code.
+ */
+
+use Neos\Media\Domain\Model\AssetSource\AssetProxy\AssetProxy;
+use Neos\Media\Domain\Model\AssetSource\AssetSource;
+
+interface AssetSourceAwareInterface
+{
+    /**
+     * @param string $assetSourceIdentifier
+     */
+    public function setAssetSourceIdentifier(string $assetSourceIdentifier): void;
+
+    /**
+     * @return string
+     */
+    public function getAssetSourceIdentifier(): ?string;
+
+    /**
+     * @return AssetSource|null
+     */
+    public function getAssetSource(): ?AssetSource;
+
+    /**
+     * @return AssetProxy|null
+     */
+    public function getAssetProxy(): ?AssetProxy;
+}
