+++ Neos.Media/Classes/Domain/Model/AssetSource/Neos/NeosAssetProxy.php
@@ -1,8 +1,8 @@
 <?php
-namespace Neos\Media\Browser\AssetSource\Neos;
+namespace Neos\Media\Domain\Model\AssetSource\Neos;
 
 /*
- * This file is part of the Neos.Media.Browser package.
+ * This file is part of the Neos.Media package.
  *
  * (c) Contributors of the Neos Project - www.neos.io
   *
@@ -11,8 +11,8 @@ namespace Neos\Media\Browser\AssetSource\Neos;
  * source code.
  */
 
-use Neos\Media\Browser\AssetSource\AssetProxy\AssetProxy;
-use Neos\Media\Browser\AssetSource\AssetSource;
+use Neos\Media\Domain\Model\AssetSource\AssetProxy\AssetProxy;
+use Neos\Media\Domain\Model\AssetSource\AssetSource;
 use Neos\Flow\Annotations\Inject;
 use Neos\Flow\Http\Uri;
 use Neos\Flow\Persistence\PersistenceManagerInterface;
