+++ Neos.Media/Classes/Domain/Model/AssetSource/SupportsTagging.php
@@ -1,8 +1,8 @@
 <?php
-namespace Neos\Media\Browser\AssetSource;
+namespace Neos\Media\Domain\Model\AssetSource;
 
 /*
- * This file is part of the Neos.Media.Browser package.
+ * This file is part of the Neos.Media package.
  *
  * (c) Contributors of the Neos Project - www.neos.io
   *
